vue-sass now follows symlinks for imports
As vue-sass checked that files exists with fs.lstatSync().isFile(),
files with an symlink in their path were not detected as existing.

This fix now Checks if the path is an symlink. If so, then the symlink
will be resolved and checked if the resolved path is an exising file.